elderabot/requirements.txt
aiogram
google-generativeai
python-dotenv
Pillow
numpy
chromadb
pypdf2
aiohttp
llama-index-core